requirements.txt
albumentations
einops
fiona
geopandas
huggingface_hub
lightning==2.0.0
matplotlib
numpy
opencv-python
pandas
pillow
pyproj
rasterio
rtree
scikit-learn
scipy
seaborn
timm
torchgeo
tqdm
